requirements.txt
# telegram-multi-device-monitor Python Dependencies
# Python 3.10+ required

# Telegram Bot Framework
python-telegram-bot>=21.0,<22.0

# HTTP Client & Server
aiohttp>=3.9.0,<4.0.0
requests>=2.31.0,<3.0.0
flask>=3.0.0,<4.0.0

# Configuration & Data
pyyaml>=6.0,<7.0

# System Monitoring
psutil>=5.9.0,<6.0.0

# Optional: SSH for remote monitoring
# Uncomment if using RemoteMetricsCollector
# paramiko>=3.0.0,<4.0.0
